Add files via upload
BUSINESS USE CASE

Bismark at Luca Motors would like to get a better picture of her car sales from the last month. Using the dataset, A report is created to  shows the locations of the buyers, color-coded by the type of car purchased with a legend. Underneath the report is a  table that shows the number of sales and average sale price by car type.